First pass on OcrPostProcessor
miniOCR code i tried to adapt is not working yet. This needs anther processor in chain that passes (invokes this as child) once per page. The idea is that each processor is Atomic. So this one only deals with a single page. Too tired to keep coding. Let's see if i figure out what is missing from the XML transformation. Probably something stupid on my side.
I'm also having now with this trouble reading the value back. Its a string, but PHP on decode things its an object.. gosh