This is the largest change
If output as a file is desired a new %output file can be given. A lot of string manipulation to get the file names right back and forth and we may want to clean/generalized and also make some long code into methods we can reuse. But it actually works! WARC into WACZ without any issues!