Tiny Updates
Make sure we have the arguments we need before trying to access them
Also skip children processors if they are incompatible with the parent. Means if the child expects, e.g a filepath and the parent does not provide one.

To make this more compatible (and less akward)
i changed
FROM:
 $data->filelocation = $filelocation;
TO:
 $data->filepath = $filelocation;