HOCR for a given file from an attached file (Islandora/etc migrations)
- Very simplistic for now:
Made for Images (we can also for PDFs but that is a bit more work, ask me @patdunlavey @alliomeria if this is needed)
Will check inside as:text for a matching width/height HOCR file. (only one, not made YET for multiple, also ask me if this is needed)
IF found will get from anywhere it is stored/convert into miniOCR and skip tesseract.
Works. I'm cool. Just not cool enough to be invited for dinner!