Adds an ML Filter. Still Working on it. Does nothing yet
But it will. Basically i can't pass an Image URL and a bounding box via GET (damn Drupal 10.2) but! i can gzip and base64 encode. I will that next.
The idea here is that this filter can be fed via "ajax" by another view/formatter/plugin with a JSON structure that i will then decode here. I will check, if no bounding box was given, if i have a Vector for the same type configured in Solr and use that one, if not, i will call the corresponding API on the NLP container and the generate a vector on the fly (well like 3 seconds at least) and then alter the query to do a KNN... cool stuff here