The live backend ML vectorizing on query time is working... time to !KNN
So, now that we are actually
- Fetching an image
- Allowing a region
- calling the processor on the backend to generate vector (and object detection... will be useful in the future too)
We need to start thinking of query/alter via tags. Similar to what we do with the OCR one. But there are many things here we will have to figure out... like how does KNN interact with other ones? The prequery v/s the post query and how we will deal with facets. But baby steps first. We need to add a new query option, alter/act on it and see if !KNN works first