Small updates. Insightface is working now
One annoying thing is that because how i built Runners a runner can not generate (without a pager) multiple Flavors. Insightface so far is the one where embeddings have a closer alignment with the logic of detecting a face. The main features of a faces are encoded in the Vector which in test allowed me to even get family members (tested with myself and mom!) but also means that if a single image has multiple detections i need multiple Vectors, and Solr allows on vector per field/per document. Will explore the one to many option in our code @alliomeria